[9.5.0] add Item no found exception (#8)
* fix no java

* aaa

* create a structure for register

* aaa

* [9.50] update item management

allow users to upload items, get all unsold items, get items by id, update item info, and delete item

* Update README.md

* add some exception handle

+++ SecondHandMarketBackend/src/main/java/com/bearbazzar/secondhandmarketbackend/service/ItemService.java
@@ -40,6 +40,9 @@ public class ItemService {
           return null;
     }
     public void deleteItem(Long id) {
+        if(!itemRepository.existsById(id)) {
+            throw new ItemNoExistException("Item does not exists");
+        }
           itemRepository.deleteById(id);
     }
 }
